refactor(auth): abstracted ApolloServer config
